Move `tests` to `tests/web`
