+++ e2e/tests/web/with-ssh-node/ssh.spec.ts-snapshots/ls-output-with-ssh-node-darwin.png
